RF: Move fx into `get`; not tested separately; mark as internal helper
